requirements.txt
feedparser==6.0.10
newspaper==0.1.0.7
newspaper3k==0.2.8
Pillow==9.1.1
requests==2.28.0
